refactor: move get_cursor_position to adapters
