+++ crates/pgt_lsp/src/handlers.rs
@@ -1,4 +1,3 @@
 pub(crate) mod code_actions;
 pub(crate) mod completions;
-mod helper;
 pub(crate) mod text_document;
